Fix bug: close #40

+++ config_reader.py
@@ -49,9 +49,10 @@ class ConfigReader:
         for key, val_type in self.__get_required_config_settings():
             if key not in data:
                 raise ConfigReaderError("12-002", "KeyError - W pliku config.json nie ma: " + key)
-            if val_type == "path":
+            if val_type == "path_dir" or val_type == "path_file":
+                is_file = (val_type == "path_file")
                 data[key] = os.path.expandvars(data[key])
-                self.__check_path(data[key])
+                self.__check_path(data[key], is_file)
         self.__on_add_log(2, "CNF_READ", "", "Pobrano konfigurację", False)
         return data
 
@@ -66,28 +67,29 @@ class ConfigReader:
             ["minimum_number_of_lines_to_write_in_log_file", "int"],
             ["socket_timeout", "float"],
             ["number_of_lanes", "int"],
-            ["dir_fonts", "path"],
-            ["dir_template_lane", "path"],
-            ["dir_template_main", "path"],
-            ["file_output_lane", "path"],
-            ["file_output_main", "path"],
-            ["dir_instructions_lane", "path"],
-            ["dir_instructions_main", "path"],
-            ["file_with_licenses_config", "path"],
-            ["file_with_category_types", "path"],
-            ["file_with_game_types", "path"],
+            ["dir_fonts", "path_dir"],
+            ["dir_template_lane", "path_dir"],
+            ["dir_template_main", "path_dir"],
+            ["file_output_lane", "path_file"],
+            ["file_output_main", "path_file"],
+            ["dir_instructions_lane", "path_dir"],
+            ["dir_instructions_main", "path_dir"],
+            ["file_with_licenses_config", "path_file"],
+            ["file_with_category_types", "path_file"],
+            ["file_with_game_types", "path_file"],
             ["loop_interval_ms", "int"],
             ["server_ip", "str"],
             ["server_port", "str"]
+            ["server_port", "str"],
         ]
         return list_settings
 
-    def __check_path(self, path: str) -> None:
+    def __check_path(self, path: str, is_file: bool) -> None:
         if not os.path.exists(path):
             os.makedirs(os.path.dirname(path), exist_ok=True)
-
-            with open(path, "w") as f:
-                pass
+            if is_file:
+                with open(path, "w") as f:
+                    pass
             self.__on_add_log(7, "CNF_PATH_NOTEXISTED", "", f"Nie istniałą ścieżka: {path}", True)
         else:
             self.__on_add_log(1, "CNF_PATH_OK", "", f"Ścieżka poprawna: {path}", True)
